feat: slice appointments array data to two indices

+++ hooks/dateUtility.ts
@@ -1,7 +1,7 @@
 export function formatDate(dateString: string) {
   const date = new Date(dateString);
   function getMonthName(): string {
-    const options: Intl.DateTimeFormatOptions = { month: "long" };
+    const options: Intl.DateTimeFormatOptions = { month: "short" };
     return new Intl.DateTimeFormat("en-US", options).format(date);
   }
   function getDayNumber(): number {
